+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 ### Fixed
 - Fix some missing translations.
+- Fix a bug with Window capture mode not correctly setting up the scene.
 
 ## [6.5.0] - 2025-01-03
 ### Changed